requirements.txt
requests
beautifulsoup4
urllib3
numpy
pandas
psycopg2
python-dotenv
pandas
lxml
apache-airflow-providers-postgres
logging
matplotlib
seaborn
scikit-learn
statsmodels
tqdm
xgboost
apache-airflow
